moved searchbar from navbar into seperate component, and moved avatarUrl into util.js.
