+++ frontend/src/lib/components/navbar/navbar.svelte
@@ -1,71 +1,25 @@
 <script>
   import { goto } from '$app/navigation';
-  import { page } from '$app/stores';
-  import { blur } from 'svelte/transition';
   
   import { toast } from 'svelte-sonner';
 
-  import { Cog, CookingPot, LogIn, LogOut, LoaderCircle, CircleUser, Menu, Search } from 'lucide-svelte';
-  import { Input } from '$lib/components/ui/input/index.js';
+  import { Cog, CookingPot, LogIn, LogOut, CircleUser, Menu, ClipboardPen } from 'lucide-svelte';
   import { Button } from '$lib/components/ui/button/index.js';
   import * as DropdownMenu from '$lib/components/ui/dropdown-menu/index.js';
   import * as Sheet from '$lib/components/ui/sheet/index.js';
-  import Separator from '../ui/separator/separator.svelte';
-  import Label from '../ui/label/label.svelte';
   
   import Navlink from './NavLink.svelte';
   import ThemeToggle from '../ThemeToggle/ThemeToggle.svelte';
+  import SearchBar from '../SerachBar/SearchBar.svelte';
 
   import { updateAuthState, isAuthenticated, user } from '../../../stores/authStore.js';
   import { avatarStore } from '../../../stores/avatarStore.js';
 
   import authApi from '$lib/api/authApi.js';
-  import userApi from '$lib/api/userApi';
-  import recipeApi from '$lib/api/recipeApi';
   
+  import { avatarUrl } from '$lib/utils/util.js';
 
-  let searchValue = $state("");
-  let searchFocused = $state(false);
-  let searchResults = $state([]);
-  let recipeSearchResults = $state([]);
-  let userSearchResults = $state([]);
-  let searchLoading = $state(false);
 
-  let debounceTimeout;
-
-  $effect(() => {
-    if (searchValue) {
-      searchLoading = true;
-      userSearchResults = [];
-      recipeSearchResults = [];
-      clearTimeout(debounceTimeout);
-
-      debounceTimeout = setTimeout(async () => {
-        searchLoading = true;
-        try {
-          // Run searches in parallel
-          await Promise.all([
-            handleUserSearch(searchValue),
-            handleRecipeSearch(searchValue),
-          ]);
-        } catch(error) {
-
-        } finally {
-          // Ensure loading is set to false even if there are errors
-          searchLoading = false;
-        }
-      }, 300); // 300ms debounce delay
-    } else {
-      userSearchResults = [];
-      recipeSearchResults = [];
-    }
-  });
-
-  const avatarUrl = (userId) => {
-    return import.meta.env.VITE_API_URL
-      ? `${import.meta.env.VITE_API_URL}/users/${userId}/avatar`
-      : `/api/users/${userId}/avatar`;
-  };
 
   const handleLogout = async () => {
     try {
@@ -76,30 +30,6 @@
 
     }catch(error) {
       toast.error(error.message);
-    }
-  };
-
-
-  const handleUserSearch = async (query) => {
-    if (query.length > 2) {
-      try {
-        const results = await userApi.getUsersByPartialUsername(query);
-        userSearchResults = results;
-      } catch (error) {
-      }
-      
-    } else {
-      userSearchResults = [];
-    }
-  };
-
-  const handleRecipeSearch = async (query) => {
-    if (query.length > 2) {
-      const results = await recipeApi.getRecipesByPartialName(query);
-      recipeSearchResults = results;
-
-    } else {
-      recipeSearchResults = [];
     }
   };
 
@@ -114,9 +44,7 @@
 
 <header class="bg-background sticky top-0 border-b">
   <div class="mx-auto max-w-7xl flex h-16 items-center gap-4 px-4 md:px-6">
-    <nav
-      class="hidden flex-col gap-6 md:flex md:flex-row md:items-center md:gap-5 lg:gap-6"
-    >
+    <nav class="hidden flex-col gap-6 md:flex md:flex-row md:items-center md:gap-5 lg:gap-6">
       <a
         href={$isAuthenticated ? "/explore" : "/"}
         class="flex items-center gap-2 text-lg font-semibold md:text-base"
@@ -160,200 +88,20 @@
     </Sheet.Root>
     
     <div class="flex w-full items-center gap-4 md:ml-auto md:gap-2 lg:gap-4">
-      <div class="ml-auto flex-1 sm:flex-initial">
+      <div class="ml-auto flex items-center sm:flex-initial">
+
+        <SearchBar/>
+      
         {#if !$isAuthenticated}
-          <Navlink exact={true} href="/login"
-            >Login <LogIn class="h-5 w-5 inline-block" /></Navlink
-          >
+          <Navlink className="ml-4" exact={true} href="/login">
+            Login <LogIn class="h-5 w-5 inline-block" />
+          </Navlink>
+
+          <a class="ml-4 px-4 py-2 bg-primary text-primary-foreground hover:bg-primary/90 rounded-xl font-semibold" exact={true} href="/register">
+            Register <ClipboardPen class="h-5 w-5 inline-block"/>
+          </a>
         {/if}
       </div>
-
-      <!-- SEARCH START -->
-      <!-- {#if $isAuthenticated} -->
-        <div class="relative z-50">
-          <Button
-            variant="ghost"
-            size="sm"
-            class="rounded-lg flex items-center gap-2 hover:bg-gray-100 dark:hover:bg-slate-800"
-            onclick={() => (searchFocused = true)}
-          >
-            <Search class="h-4 w-4" />
-            <span>Search</span>
-          </Button>
-
-          <!-- Search results dropdown -->
-          {#if searchFocused}
-            <!-- Blocker to disable all interaction with the rest of the page -->
-            <div
-              class="fixed inset-0 z-40"
-              style="background: transparent;"
-              tabindex="-1"
-              role="button"
-              onclick={() => (searchFocused = false)}
-              onkeydown={(e) => e.key === "Escape" && (searchFocused = false)}
-              aria-label="Close search"
-            ></div>
-
-            <div
-              transition:blur={{ duration: 250 }}
-              class="fixed z-50 inset-0 flex items-start justify-center pt-16 bg-black/40"
-              style="pointer-events: none;"
-            >
-              <div
-                class="w-full max-w-xl bg-white dark:bg-slate-900 border border-gray-200 dark:border-gray-700 rounded-lg shadow-xl z-50 overflow-hidden mt-10"
-                style="pointer-events: auto;"
-              >
-                <!-- Panel header with search input -->
-                <div
-                  class="px-4 py-3 border-b border-gray-200 dark:border-gray-700 flex items-center justify-between bg-gray-50 dark:bg-slate-800"
-                >
-                  <div class="flex-grow flex items-center gap-2">
-                    <Search
-                      class="h-4 w-4 text-gray-500 dark:text-gray-400 flex-shrink-0"
-                    />
-                    <Input
-                      type="search"
-                      placeholder="Search..."
-                      class="border-0 shadow-none focus-visible:ring-0 focus-visible:ring-offset-0 h-9 bg-transparent"
-                      bind:value={searchValue}
-                      autofocus
-                    />
-                  </div>
-                </div>
-
-                <!-- Scrollable content -->
-                <div class="max-h-[60vh] overflow-y-auto">
-                  {#if recipeSearchResults.length}
-                    <div class="px-4 py-3">
-                      <h3
-                        class="text-sm font-medium text-gray-600 dark:text-gray-400 uppercase tracking-wider flex items-center gap-2"
-                      >
-                        <CookingPot class="h-4 w-4" />
-                        Recipes
-                      </h3>
-                      <ul
-                        class="mt-2 divide-y divide-gray-100 dark:divide-gray-800"
-                      >
-                        {#each recipeSearchResults as recipe}
-                          <button
-                            onclick={() => handleNavigate(recipe.id)}
-                            class="w-full flex items-center gap-3 p-3 hover:bg-gray-50 dark:hover:bg-gray-800 cursor-pointer rounded transition-colors duration-150"
-                          >
-                            <img
-                              src={recipe.image ||
-                                "/recipe-image-placeholder.png"}
-                              alt={recipe.name}
-                              class="h-14 w-14 rounded-md object-cover shadow-sm border border-gray-200 dark:border-gray-700"
-                            />
-                            <div class="flex-1 min-w-0">
-                              <p
-                                class="text-sm font-medium text-gray-900 dark:text-gray-100 truncate text-left"
-                              >
-                                {recipe.name}
-                              </p>
-                              {#if recipe.description}
-                                <p
-                                  class="text-xs text-gray-500 dark:text-gray-400 line-clamp-2 text-left"
-                                >
-                                  {recipe.description}
-                                </p>
-                              {/if}
-                            </div>
-                          </button>
-                        {/each}
-                      </ul>
-                    </div>
-                  {/if}
-
-                  {#if userSearchResults.length}
-                    <div
-                      class="px-4 py-3 {recipeSearchResults.length
-                        ? 'border-t border-gray-200 dark:border-gray-700'
-                        : ''}"
-                    >
-                      <h3
-                        class="text-sm font-medium text-gray-600 dark:text-gray-400 uppercase tracking-wider flex items-center gap-2"
-                      >
-                        <CircleUser class="h-4 w-4" />
-                        Users
-                      </h3>
-                      <ul
-                        class="mt-2 divide-y divide-gray-100 dark:divide-gray-800"
-                      >
-                        {#each userSearchResults as usr}
-                          <button
-                            onclick={() => {
-                              goto(`/users/${usr.id}`);
-                              searchValue = "";
-                              searchFocused = false;
-                            }}
-                            class="w-full flex items-center gap-3 p-3 hover:bg-gray-50 dark:hover:bg-gray-800 cursor-pointer rounded transition-colors duration-150"
-                          >
-                            <div class="flex-shrink-0">
-                              {#if usr.avatar}
-                                <img
-                                  src={avatarUrl(usr.id)}
-                                  alt={usr.username}
-                                  class="h-10 w-10 rounded-full object-cover border border-gray-200 dark:border-gray-700 shadow-sm"
-                                />
-                              
-                              {:else}
-                                <CircleUser class="h-10 w-10"/>
-                              {/if}
-                            </div>
-                            <div class="flex-1 min-w-0">
-                              <p
-                                class="text-left text-sm font-medium text-gray-900 dark:text-gray-100"
-                              >
-                                {usr.username}
-                              </p>
-                            </div>
-                          </button>
-                        {/each}
-                      </ul>
-                    </div>
-                  {/if}
-
-                  {#if searchLoading}
-                    <div class="py-10 px-8 text-center">
-                      <div
-                        class="inline-flex items-center justify-center w-14 h-14 mb-4"
-                      >
-                        <LoaderCircle
-                          class="h-6 w-6 animate-spin text-gray-500 dark:text-gray-400"
-                        />
-                      </div>
-                    </div>
-                  {/if}
-
-                  {#if !searchLoading && !recipeSearchResults.length && !userSearchResults.length}
-                    <div class="py-10 px-8 text-center">
-                      <div
-                        class="inline-flex items-center justify-center w-14 h-14 mb-4"
-                      >
-                        <Search
-                          class="h-6 w-6 text-gray-500 dark:text-gray-400"
-                        />
-                      </div>
-                      {#if searchValue}
-                        <p class="text-gray-700 dark:text-gray-300 font-medium">
-                          No results found for "{searchValue}"
-                        </p>
-
-                        {:else}
-                          <p class="text-gray-700 dark:text-gray-300 font-medium">
-                            Type to search for delicious recipes or discover other users...
-                          </p>
-                      {/if}
-                    </div>
-                  {/if}
-                </div>
-              </div>
-            </div>
-          {/if}
-        </div>
-        <!-- END SEARCH INPUT AND DROPDOWN -->
-      <!-- {/if} -->
 
       {#if $isAuthenticated && $user}
       <DropdownMenu.Root>
